requirements.txt
pandas==2.0.3
numpy==1.24.3
flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
pycryptodome==3.18.0
nltk==3.8.1
scikit-learn==1.3.0
matplotlib==3.7.2
openpyxl==3.1.2
reportlab==4.0.4
requests==2.31.0
